Extract RedirectLimitReached to its own file
